LayoutTests/media/media-source/media-source-real-duration-after-append-expected.txt


EVENT(sourceopen)
EVENT(updateend)
EXPECTED (video.duration == '0') OK
EVENT(updateend)
EXPECTED (video.duration == '10') OK
END OF TEST
